// ClientApp_old/src/services/openapi/api.ts
export interface AddFoodDiaryRequest {
    /**
     * 
     * @type {string}
     * @memberof AddFoodDiaryRequest
     */
    'foodId': string;
    /**
     * 
     * @type {MealTypes}
     * @memberof AddFoodDiaryRequest
     */
    'mealType': MealTypes;
    /**
     * 
     * @type {string}
     * @memberof AddFoodDiaryRequest
     */
    'servingSizeId': string;
    /**
     * 
     * @type {number}
     * @memberof AddFoodDiaryRequest
     */
    'quantity': number;
    /**
     * 
     * @type {string}
     * @memberof AddFoodDiaryRequest
     */
    'entryDate': string;
}
// ...
export const DiaryType = {
    FoodDiary: 'FoodDiary',
    RecipeDiary: 'RecipeDiary'
} as const;

export type DiaryType = typeof DiaryType[keyof typeof DiaryType];


/**
 * 
 * @export
 * @interface Energy
 */
export interface Energy {
    /**
     * 
     * @type {string}
     * @memberof Energy
     */
    'unit': string;
    /**
     * 
     * @type {number}
     * @memberof Energy
     */
    'value': number;
}
// ...
export interface FoodDto {
    /**
     * 
     * @type {string}
     * @memberof FoodDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof FoodDto
     */
    'type': string;
    /**
     * 
     * @type {string}
     * @memberof FoodDto
     */
    'brandName': string;
    /**
     * 
     * @type {string}
     * @memberof FoodDto
     */
    'countryCode': string;
    /**
     * 
     * @type {string}
     * @memberof FoodDto
     */
    'name': string;
    /**
     * 
     * @type {NutritionalContent}
     * @memberof FoodDto
     */
    'nutritionalContents': NutritionalContent;
    /**
     * 
     * @type {{ [key: string]: ServingSizeDto; }}
     * @memberof FoodDto
     */
    'servingSizes': { [key: string]: ServingSizeDto; };
}
// ...
export const MealTypes = {
    Breakfast: 'Breakfast',
    Lunch: 'Lunch',
    Dinner: 'Dinner',
    Snacks: 'Snacks'
} as const;

export type MealTypes = typeof MealTypes[keyof typeof MealTypes];
// ...
export interface NutritionalContent {
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'calcium': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'carbohydrates': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'cholesterol': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'fat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'fiber': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'iron': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'monounsaturatedFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'netCarbs': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'polyunsaturatedFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'potassium': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'protein': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'saturatedFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'sodium': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'sugar': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'transFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'vitaminA': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'vitaminC': number;
    /**
     * 
     * @type {Energy}
     * @memberof NutritionalContent
     */
    'energy': Energy;
}
// ...
export interface ServingSizeDto {
    /**
     * 
     * @type {string}
     * @memberof ServingSizeDto
     */
    'id': string;
    /**
     * 
     * @type {number}
     * @memberof ServingSizeDto
     */
    'nutritionMultiplier': number;
    /**
     * 
     * @type {number}
     * @memberof ServingSizeDto
     */
    'value': number;
    /**
     * 
     * @type {string}
     * @memberof ServingSizeDto
     */
    'unit': string;
}
// ...
export interface UpdateFoodDiaryRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateFoodDiaryRequest
     */
    'id': string;
    /**
     * 
     * @type {number}
     * @memberof UpdateFoodDiaryRequest
     */
    'quantity': number;
    /**
     * 
     * @type {string}
     * @memberof UpdateFoodDiaryRequest
     */
    'servingSizeId': string;
    /**
     * 
     * @type {MealTypes}
     * @memberof UpdateFoodDiaryRequest
     */
    'mealType': MealTypes;
}
// ...
export class FoodDiariesApi extends BaseAPI {
    /**
     * 
     * @param {AddFoodDiaryRequest} addFoodDiaryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FoodDiariesApi
     */
    public addFoodDiary(addFoodDiaryRequest: AddFoodDiaryRequest, options?: RawAxiosRequestConfig) {
        return FoodDiariesApiFp(this.configuration).addFoodDiary(addFoodDiaryRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string | null} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FoodDiariesApi
     */
    public deleteFoodDiary(id: string | null, options?: RawAxiosRequestConfig) {
        return FoodDiariesApiFp(this.configuration).deleteFoodDiary(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string | null} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FoodDiariesApi
     */
    public getFoodDiaryById(id: string | null, options?: RawAxiosRequestConfig) {
        return FoodDiariesApiFp(this.configuration).getFoodDiaryById(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {UpdateFoodDiaryRequest} updateFoodDiaryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FoodDiariesApi
     */
    public updateFoodDiary(updateFoodDiaryRequest: UpdateFoodDiaryRequest, options?: RawAxiosRequestConfig) {
        return FoodDiariesApiFp(this.configuration).updateFoodDiary(updateFoodDiaryRequest, options).then((request) => request(this.axios, this.basePath));
    }
}